+++ src/main/java/com/openstudies/hibernate/services/courses/CourseService.java
@@ -10,6 +10,6 @@ public interface CourseService {
     public void update(Course course);
     public void delete(Long id);
     public void getCourseByName(String courseName);
-    public List<Course> getOwnerCourses(Long id);
+    public List<Course> getOwnerCourses(Integer id);
 
 }
